// apps/web/src/lib/api/generated/model/activityInKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export type ActivityInKind = typeof ActivityInKind[keyof typeof ActivityInKind];


export const ActivityInKind = {
  note: 'note',
  call: 'call',
  meeting: 'meeting',
  whatsapp: 'whatsapp',
  sms: 'sms',
  email: 'email',
  visit: 'visit',
} as const;