// apps/web/src/lib/api/generated/model/clientOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { ClientOutAddress } from './clientOutAddress';

export interface ClientOut {
  address: ClientOutAddress;
  alt_phone: string | null;
  archived_at: string | null;
  company_name: string | null;
  created_at: string;
  date_of_birth: string | null;
  display_name: string;
  email: string | null;
  first_name: string | null;
  gender: string | null;
  household_id: string | null;
  household_role: string | null;
  id: string;
  id_number_hint: string | null;
  id_type: string | null;
  kind: string;
  kra_pin: string | null;
  last_name: string | null;
  marketing_consent: boolean;
  marketing_consent_at: string | null;
  notes: string | null;
  occupation: string | null;
  other_names: string | null;
  owner_user_id: string | null;
  phone: string | null;
  preferred_channel: string;
  referred_by_id: string | null;
  source: string | null;
  status: string;
  tags: string[];
  updated_at: string;
  version: number;
}