// apps/web/src/lib/api/generated/model/documentCreateCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export type DocumentCreateCategory = typeof DocumentCreateCategory[keyof typeof DocumentCreateCategory];


export const DocumentCreateCategory = {
  kyc_id: 'kyc_id',
  kyc_pin: 'kyc_pin',
  kyc_other: 'kyc_other',
  policy_schedule: 'policy_schedule',
  certificate: 'certificate',
  vehicle_logbook: 'vehicle_logbook',
  claim: 'claim',
  generated: 'generated',
  branding: 'branding',
  other: 'other',
} as const;